+++ torch_geometric/nn/pool/glob.py
@@ -1,7 +1,8 @@
 from typing import Optional
 
 from torch import Tensor
-from torch_scatter import scatter
+
+from torch_geometric.utils import scatter
 
 
 def global_add_pool(x: Tensor, batch: Optional[Tensor],
@@ -28,7 +29,7 @@ def global_add_pool(x: Tensor, batch: Optional[Tensor],
     if batch is None:
         return x.sum(dim=-2, keepdim=x.dim() == 2)
     size = int(batch.max().item() + 1) if size is None else size
-    return scatter(x, batch, dim=-2, dim_size=size, reduce='add')
+    return scatter(x, batch, dim=-2, dim_size=size, reduce='sum')
 
 
 def global_mean_pool(x: Tensor, batch: Optional[Tensor],