+++ torch_geometric/utils/loop.py
@@ -2,9 +2,9 @@ from typing import Optional, Tuple, Union
 
 import torch
 from torch import Tensor
-from torch_scatter import scatter
 
 from torch_geometric.typing import OptTensor
+from torch_geometric.utils import scatter
 
 from .num_nodes import maybe_num_nodes